feat(shopping): move shopping-list add flow to a browser side panel
Alternative to #251's popup-embedded UI and #253's dedicated page.
Moves the search -> pick recipe -> pick list -> add flow to a side
panel (entrypoints/sidepanel/), which WXT wires automatically per
browser: Chrome MV3 gets side_panel.default_path (+ the auto-added
sidePanel permission), Firefox gets sidebar_action. Unlike the
dedicated-page variant, the panel stays open alongside the page
you're browsing instead of living in its own tab.

The popup's "Open Shopping List" button opens the panel via
chrome.sidePanel.open() on Chrome or sidebarAction.open() on Firefox,
called synchronously in the click handler (both APIs require an
unbroken user-gesture call stack).

Manifest impact: adds the "sidePanel" permission on Chrome only
(auto-added by WXT for any entrypoints/sidepanel/ page) and a
sidebar_action key on Firefox. No host permission changes.
