+++ entrypoints/popup/App.tsx
@@ -512,10 +512,7 @@ function App() {
                             </label>
                         </div>
 
-                        <ShoppingListQuickAdd
-                            mealieServer={mealieServer}
-                            mealieApiToken={mealieApiToken}
-                        />
+                        <ShoppingListPanelButton />
 
                         <ActivityLog />
 
@@ -536,203 +533,33 @@ function App() {
     );
 }
 
-type ShoppingStatus = { type: 'success' | 'error'; message: string };
+/**
+ * Opens the shopping-list side panel. Must run synchronously inside the click handler
+ * (no `await` before it) — both Chrome's `sidePanel.open()` and Firefox's
+ * `sidebarAction.open()` require an unbroken user-gesture call stack, and lose that
+ * association across a microtask boundary. `WINDOW_ID_CURRENT` lets Chrome open the
+ * panel for "this" window without an async `chrome.windows`/`chrome.tabs` lookup first.
+ */
+function openShoppingListPanel() {
+    if (chrome.sidePanel?.open) {
+        void chrome.sidePanel.open({ windowId: chrome.windows.WINDOW_ID_CURRENT });
+        return;
+    }
 
-function ShoppingListQuickAdd({
-    mealieServer,
-    mealieApiToken,
-}: {
-    mealieServer: string;
-    mealieApiToken: string;
-}) {
-    const [isOpen, setIsOpen] = useState(false);
-    const [query, setQuery] = useState('');
-    const [searching, setSearching] = useState(false);
-    const [searched, setSearched] = useState(false);
-    const [results, setResults] = useState<RecipeSummary[]>([]);
-    const [selectedRecipe, setSelectedRecipe] = useState<RecipeSummary | null>(null);
-    const [lists, setLists] = useState<ShoppingListSummary[] | null>(null);
-    const [selectedListId, setSelectedListId] = useState('');
-    const [loadingLists, setLoadingLists] = useState(false);
-    const [adding, setAdding] = useState(false);
-    const [status, setStatus] = useState<ShoppingStatus | null>(null);
+    // Firefox's sidebar_action equivalent — not declared in @types/chrome.
+    const sidebarAction = (chrome as unknown as { sidebarAction?: { open: () => Promise<void> } })
+        .sidebarAction;
+    if (sidebarAction?.open) {
+        void sidebarAction.open();
+    }
+}
 
-    const resetSelection = () => {
-        setSelectedRecipe(null);
-        setStatus(null);
-    };
-
-    const handleSearch = async () => {
-        const trimmed = query.trim();
-        if (!trimmed) return;
-
-        setSearching(true);
-        setSearched(false);
-        setStatus(null);
-        resetSelection();
-
-        const matches = await searchRecipesByName(trimmed, mealieServer, mealieApiToken);
-
-        setResults(matches);
-        setSearched(true);
-        setSearching(false);
-    };
-
-    const handleSelectRecipe = async (recipe: RecipeSummary) => {
-        setSelectedRecipe(recipe);
-        setStatus(null);
-
-        if (lists) return;
-
-        setLoadingLists(true);
-        const fetched = await getShoppingLists(mealieServer, mealieApiToken);
-        setLoadingLists(false);
-
-        if (fetched === 'failure') {
-            setStatus({ type: 'error', message: 'Could not load shopping lists.' });
-            return;
-        }
-
-        setLists(fetched);
-        const [firstList] = fetched;
-        if (firstList) {
-            setSelectedListId(firstList.id);
-        }
-    };
-
-    const handleAddToList = async () => {
-        if (!selectedRecipe || !selectedListId) return;
-
-        setAdding(true);
-        setStatus(null);
-        const success = await addRecipeToShoppingList(
-            selectedListId,
-            selectedRecipe.id,
-            mealieServer,
-            mealieApiToken,
-        );
-        setAdding(false);
-
-        const listName = lists?.find((list) => list.id === selectedListId)?.name ?? 'your list';
-        setStatus(
-            success
-                ? { type: 'success', message: `Added "${selectedRecipe.name}" to ${listName}.` }
-                : {
-                      type: 'error',
-                      message: 'Failed to add ingredients. Check the activity log for details.',
-                  },
-        );
-    };
-
+function ShoppingListPanelButton() {
     return (
-        <div className="collapsible-section">
-            <button
-                type="button"
-                className="collapsible-toggle"
-                onClick={() => setIsOpen((prev) => !prev)}
-                aria-expanded={isOpen}
-            >
-                🛒 Add Recipe to Shopping List {isOpen ? '▲' : '▼'}
+        <div className="activity-log">
+            <button className="activity-log-link" onClick={openShoppingListPanel}>
+                🛒 Open Shopping List ↗
             </button>
-            {isOpen && (
-                <div className="collapsible-body">
-                    <div className="shopping-search-row">
-                        <input
-                            type="text"
-                            placeholder="Search your recipes…"
-                            value={query}
-                            onChange={(e) => setQuery(e.target.value)}
-                            onKeyDown={(e) => {
-                                if (e.key === 'Enter' && !searching && query.trim()) {
-                                    void handleSearch();
-                                }
-                            }}
-                        />
-                        <button
-                            type="button"
-                            onClick={() => void handleSearch()}
-                            disabled={searching || !query.trim()}
-                        >
-                            {searching ? 'Searching…' : 'Search'}
-                        </button>
-                    </div>
-
-                    {searched && results.length === 0 && (
-                        <p className="shopping-empty">
-                            No recipes found for &quot;{query.trim()}&quot;.
-                        </p>
-                    )}
-
-                    {results.length > 0 && !selectedRecipe && (
-                        <ul className="shopping-results">
-                            {results.map((recipe) => (
-                                <li key={recipe.id}>
-                                    <button
-                                        type="button"
-                                        onClick={() => void handleSelectRecipe(recipe)}
-                                    >
-                                        {recipe.name}
-                                    </button>
-                                </li>
-                            ))}
-                        </ul>
-                    )}
-
-                    {selectedRecipe && (
-                        <div className="shopping-selected">
-                            <p>
-                                Adding <strong>{selectedRecipe.name}</strong>
-                            </p>
-                            {loadingLists && <p>Loading your shopping lists…</p>}
-                            {!loadingLists && lists && lists.length === 0 && (
-                                <p className="shopping-empty">
-                                    No shopping lists found — create one in Mealie first.
-                                </p>
-                            )}
-                            {!loadingLists && lists && lists.length > 0 && (
-                                <>
-                                    <select
-                                        value={selectedListId}
-                                        onChange={(e) => setSelectedListId(e.target.value)}
-                                    >
-                                        {lists.map((list) => (
-                                            <option key={list.id} value={list.id}>
-                                                {list.name ?? 'Untitled list'}
-                                            </option>
-                                        ))}
-                                    </select>
-                                    <button
-                                        type="button"
-                                        onClick={() => void handleAddToList()}
-                                        disabled={adding}
-                                    >
-                                        {adding ? 'Adding…' : 'Add to List'}
-                                    </button>
-                                </>
-                            )}
-                            <button
-                                type="button"
-                                className="shopping-back"
-                                onClick={resetSelection}
-                            >
-                                ← Back to results
-                            </button>
-                        </div>
-                    )}
-
-                    {status && (
-                        <p
-                            className={
-                                status.type === 'success'
-                                    ? 'shopping-success'
-                                    : 'connect-error-detail'
-                            }
-                        >
-                            {status.message}
-                        </p>
-                    )}
-                </div>
-            )}
         </div>
     );
 }
