+++ entrypoints/sidepanel/style.css
@@ -1,0 +1,179 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    margin: 0;
+    font-family:
+        -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
+    background: #1a1a1a;
+    color: #e0e0e0;
+}
+
+.sidepanel {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    padding: 1rem;
+}
+
+.sidepanel-header {
+    margin-bottom: 1rem;
+}
+
+.sidepanel-header h1 {
+    margin: 0;
+    font-size: 1.2rem;
+    color: #ec7e19;
+}
+
+.sidepanel-empty {
+    color: #888;
+    text-align: center;
+    padding: 2rem 0.5rem;
+    font-size: 0.9rem;
+}
+
+.shopping-search-row {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.shopping-search-row input[type='text'] {
+    padding: 0.6rem 0.7rem;
+    border-radius: 4px;
+    border: 1px solid #444;
+    background-color: #2c2c2c;
+    color: #e0e0e0;
+    font-size: 0.95rem;
+}
+
+.shopping-search-row button {
+    padding: 0.55rem 0.8rem;
+    background: #ec7e19;
+    border: none;
+    border-radius: 4px;
+    color: white;
+    cursor: pointer;
+    font-size: 0.9rem;
+}
+
+.shopping-search-row button:hover:not(:disabled) {
+    background: #d66f15;
+}
+
+.shopping-search-row button:disabled {
+    background: #a65726;
+    color: rgba(255, 255, 255, 0.8);
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+
+.shopping-results {
+    list-style: none;
+    margin: 1rem 0 0;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+    overflow-y: auto;
+    border: 1px solid #333;
+    border-radius: 6px;
+}
+
+.shopping-results li:not(:first-child) {
+    border-top: 1px solid #333;
+}
+
+.shopping-results button {
+    width: 100%;
+    text-align: left;
+    font-size: 0.88rem;
+    padding: 0.6rem 0.7rem;
+    background-color: #222;
+    color: #e0e0e0;
+    border: none;
+    cursor: pointer;
+}
+
+.shopping-results button:hover {
+    background-color: rgba(236, 126, 25, 0.18);
+}
+
+.shopping-selected {
+    margin-top: 1rem;
+    padding: 0.9rem;
+    background-color: #0d0d0d;
+    border: 1px solid #333;
+    border-radius: 6px;
+}
+
+.shopping-selected p {
+    margin-top: 0;
+    font-size: 0.92rem;
+}
+
+.shopping-selected select {
+    width: 100%;
+    padding: 0.55rem;
+    margin-bottom: 0.6rem;
+    border-radius: 4px;
+    border: 1px solid #444;
+    background-color: #2c2c2c;
+    color: #e0e0e0;
+    font-size: 0.9rem;
+}
+
+.shopping-selected select + button {
+    width: 100%;
+    padding: 0.55rem;
+    background: #ec7e19;
+    border: none;
+    border-radius: 4px;
+    color: white;
+    cursor: pointer;
+    font-size: 0.9rem;
+}
+
+.shopping-selected select + button:hover:not(:disabled) {
+    background: #d66f15;
+}
+
+.shopping-selected select + button:disabled {
+    background: #a65726;
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+
+.shopping-back {
+    display: block;
+    margin-top: 0.5rem;
+    background: transparent;
+    border: none;
+    color: #ec7e19;
+    cursor: pointer;
+    padding: 0.3rem 0;
+    font-size: 0.85rem;
+}
+
+.shopping-back:hover {
+    text-decoration: underline;
+}
+
+.shopping-empty {
+    font-size: 0.85rem;
+    color: #888;
+    margin: 1rem 0 0;
+}
+
+.shopping-error {
+    font-size: 0.85rem;
+    color: #f0c4a8;
+    margin: 1rem 0 0;
+}
+
+.shopping-success {
+    font-size: 0.85rem;
+    color: #8fd694;
+    margin: 1rem 0 0;
+}
